+++ windows/src/engine/keyman32/preservedkeymap.cpp
@@ -234,6 +234,16 @@ BOOL PreservedKeyMap::MapKeyboard(KEYBOARD *pKeyboard, PreservedKey **pPreserved
 
   m_BaseKeyboardUsesAltGr = KeyboardGivesCtrlRAltForRAlt();   // I4592
 
+  // This is not the same as m_BaseKeyboardUsesAltGr -- we are turning 
+  // the simulation back on after (possibly) turning it off, for a 
+  // consistent experience. TODO: determine if m_BaseKeyboardUseAltGr
+  // is still needed given we always use kbdus as base for Keyman 10+
+  BOOL bSimulateAltGr = Globals::get_SimulateAltGr();
+
+  // We only want to translate RALT and RALT+SHIFT for Ctrl+Alt rules.
+  // So we exclude all our other favourite modifier keys.
+  const UINT RALT_MATCHING_MASK = TF_MOD_CONTROL | TF_MOD_ALT | TF_MOD_LCONTROL | TF_MOD_RCONTROL | TF_MOD_LALT | TF_MOD_RALT;
+
   for(i = 0; i < pKeyboard->cxGroupArray; i++)
   {
     if(pKeyboard->dpGroupArray[i].fUsingKeys)
@@ -245,6 +255,12 @@ BOOL PreservedKeyMap::MapKeyboard(KEYBOARD *pKeyboard, PreservedKey **pPreserved
   if(cKeys == 0)
   {
     return FALSE;
+  }
+
+  if (bSimulateAltGr)
+  {
+    // We might need twice as many preserved keys to map both LCtrl+LAlt+x and RAlt+x
+    cKeys *= 2;
   }
 
   if(pPreservedKeys == NULL)
@@ -267,12 +283,24 @@ BOOL PreservedKeyMap::MapKeyboard(KEYBOARD *pKeyboard, PreservedKey **pPreserved
     {
       for(j = 0; j < pGroup->cxKeyArray; j++)
       {
+        // If we have a key rule for the key, we should preserve it
         if(MapKeyRule(&pGroup->dpKeyArray[j], &pKeys[n].key))
         {
+          // Don't attempt to add the same preserved key twice. Bad things happen
           if(!IsMatchingKey(&pKeys[n], pKeys, n))
           {
             CoCreateGuid(&pKeys[n].guid);
             n++;
+
+            if (bSimulateAltGr && (pKeys[n-1].key.uModifiers & RALT_MATCHING_MASK) == TF_MOD_RALT) 
+            {
+              // Do this for RALT and RALT+SHIFT only, so we've tested against that mask
+              // Copy the key and fix modifiers
+              pKeys[n].key = pKeys[n - 1].key;
+              pKeys[n].key.uModifiers = (pKeys[n].key.uModifiers & ~TF_MOD_RALT) | TF_MOD_LCONTROL | TF_MOD_LALT;
+              CoCreateGuid(&pKeys[n].guid);
+              n++;
+            }
           }
         }
       }